Added building block view
Building block view level 1, represents the highest layer of our
system and its interactions to external systems. It has a first
decomposition of the desired program into black boxes which will be
described in lower levels.
Lower levels are to be better ordered into sub directories to avoid
a too complex md file